# apps/web/lib/api/openapi.json
{
  "components": {
    "schemas": {
...
      "BuyerLotState": {
        "description": "What the lot page needs to show a signed-in buyer's own activity on one lot.",
        "properties": {
          "contact": {
            "$ref": "#/components/schemas/ContactChoices"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "pending_hold": {
            "title": "Pending Hold",
            "type": "boolean"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "saved": {
            "title": "Saved",
            "type": "boolean"
          }
        },
        "required": [
          "saved",
          "pending_hold",
          "contact",
          "email",
          "display_name",
          "phone"
        ],
        "title": "BuyerLotState",
        "type": "object"
      },
      "ConsentChange": {
        "properties": {
          "channel": {
            "enum": [
              "email",
              "sms",
              "voice"
            ],
            "title": "Channel",
            "type": "string"
          },
          "granted": {
            "title": "Granted",
            "type": "boolean"
          },
          "tenant_id": {
            "format": "uuid",
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "tenant_id",
          "channel",
          "granted"
        ],
        "title": "ConsentChange",
        "type": "object"
      },
      "ConsentOut": {
        "description": "The current consent for one tenant and channel: its latest row.",
        "properties": {
          "channel": {
            "enum": [
              "email",
              "sms",
              "voice"
            ],
            "title": "Channel",
            "type": "string"
          },
          "granted": {
            "title": "Granted",
            "type": "boolean"
          },
          "recorded_at": {
            "format": "date-time",
            "title": "Recorded At",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "tenant_id": {
            "format": "uuid",
            "title": "Tenant Id",
            "type": "string"
          },
          "tenant_name": {
            "title": "Tenant Name",
            "type": "string"
          }
        },
        "required": [
          "tenant_id",
          "tenant_name",
          "channel",
          "granted",
          "source",
          "recorded_at"
        ],
        "title": "ConsentOut",
        "type": "object"
      },
      "ContactChoices": {
        "description": "Whether the lot's owner may contact the buyer on each channel (ADR-022). Calls (voice)\nare not offered until Phase 4.",
        "properties": {
          "email": {
            "default": false,
            "title": "Email",
            "type": "boolean"
          },
          "sms": {
            "default": false,
            "title": "Sms",
            "type": "boolean"
          }
        },
        "title": "ContactChoices",
        "type": "object"
      },
      "Created": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "title": "Created",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HoldRequestCreate": {
        "properties": {
          "contact": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContactChoices"
              },
              {
                "type": "null"
              }
            ],
            "description": "Omit to leave contact consent unchanged."
          },
          "message": {
            "default": "",
            "maxLength": 2000,
            "title": "Message",
            "type": "string"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "phone": {
            "anyOf": [
              {
                "maxLength": 32,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          }
        },
        "required": [
          "name"
        ],
        "title": "HoldRequestCreate",
        "type": "object"
      },
...
      "InquiryCreate": {
        "properties": {
          "contact": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContactChoices"
              },
              {
                "type": "null"
              }
            ],
            "description": "Signed in only. Omit to leave contact consent unchanged."
          },
          "email": {
            "anyOf": [
              {
                "format": "email",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Required when signed out; signed in, the account's is used.",
            "title": "Email"
          },
          "message": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "phone": {
            "anyOf": [
              {
                "maxLength": 32,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "turnstile_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Required when signed out.",
            "title": "Turnstile Token"
          }
        },
        "required": [
          "name",
          "message"
        ],
        "title": "InquiryCreate",
        "type": "object"
      },
...
      "LotStatus": {
        "enum": [
          "available",
          "on_hold",
          "sold"
        ],
        "title": "LotStatus",
        "type": "string"
      },
...
      "SavedLot": {
        "properties": {
          "lot_id": {
            "format": "uuid",
            "title": "Lot Id",
            "type": "string"
          },
          "number": {
            "title": "Number",
            "type": "string"
          },
          "price": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "saved_at": {
            "format": "date-time",
            "title": "Saved At",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/LotStatus"
          },
          "subdivision_name": {
            "title": "Subdivision Name",
            "type": "string"
          },
          "subdivision_slug": {
            "title": "Subdivision Slug",
            "type": "string"
          }
        },
        "required": [
          "lot_id",
          "number",
          "status",
          "price",
          "subdivision_name",
          "subdivision_slug",
          "saved_at"
        ],
        "title": "SavedLot",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Cornerpin API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/lots/{lot_id}/hold-requests": {
      "post": {
        "operationId": "create_hold_request_v1_lots__lot_id__hold_requests_post",
        "parameters": [
          {
            "in": "path",
            "name": "lot_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Lot Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "cp_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cp Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HoldRequestCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Created"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "No such published lot"
          },
          "409": {
            "description": "Not available, or already requested"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Hold Request",
        "tags": [
          "buyers"
        ]
      }
    },
    "/v1/lots/{lot_id}/inquiries": {
      "post": {
        "operationId": "create_inquiry_v1_lots__lot_id__inquiries_post",
        "parameters": [
          {
            "in": "path",
            "name": "lot_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Lot Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "cp_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cp Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InquiryCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Created"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Turnstile check failed"
          },
          "404": {
            "description": "No such published lot"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Inquiry",
        "tags": [
          "buyers"
        ]
      }
    },
...
    "/v1/me/consents": {
      "get": {
        "description": "The current answer for each owner and channel the buyer has ever answered.",
        "operationId": "my_consents_v1_me_consents_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "cp_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cp Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConsentOut"
                  },
                  "title": "Response My Consents V1 Me Consents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "My Consents",
        "tags": [
          "buyers"
        ]
      },
      "post": {
        "description": "Change an earlier answer. Only owners the buyer has already answered appear on the\naccount page, so a first answer always comes from a lot page.",
        "operationId": "change_consent_v1_me_consents_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "cp_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cp Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentChange"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "No earlier answer for that owner"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Change Consent",
        "tags": [
          "buyers"
        ]
      }
    },
    "/v1/me/lots/{lot_id}": {
      "get": {
        "operationId": "buyer_lot_state_v1_me_lots__lot_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "lot_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Lot Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "cp_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cp Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuyerLotState"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "No such published lot"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Buyer Lot State",
        "tags": [
          "buyers"
        ]
      }
    },
...
    "/v1/me/saved-lots": {
      "get": {
        "description": "Saved lots the public can still see, newest first. A lot that is unpublished drops out\nof the list but stays saved, and comes back if it is published again.",
        "operationId": "saved_lots_v1_me_saved_lots_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "cp_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cp Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SavedLot"
                  },
                  "title": "Response Saved Lots V1 Me Saved Lots Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Saved Lots",
        "tags": [
          "buyers"
        ]
      }
    },
    "/v1/me/saved-lots/{lot_id}": {
      "delete": {
        "description": "Works for any lot the buyer saved, published or not.",
        "operationId": "unsave_lot_v1_me_saved_lots__lot_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "lot_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Lot Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "cp_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cp Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unsave Lot",
        "tags": [
          "buyers"
        ]
      },
      "put": {
        "operationId": "save_lot_v1_me_saved_lots__lot_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "lot_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Lot Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "cp_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cp Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "No such published lot"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Lot",
        "tags": [
          "buyers"
        ]
      }
    },
...
  }
}